pref(intr): move page fault to a separate file
